Solve the PAT problem at GRDB level
